requirements.txt
transformers>=4.30.0
transformers[sentencepiece]>=4.30.0
accelerate>=0.26.0
huggingface_hub[hf_xet]
datasets>=2.13.0
evaluate
sacrebleu>=2.3.1
rouge_score>=0.1.2
py7zr>=0.20.4
pandas>=1.5.3
nltk>=3.8.1
tqdm>=4.65.0
PyYAML>=6.0
matplotlib>=3.7.1
torch>=2.0.1
notebook>=6.5.3
boto3>=1.26.0
mypy-boto3-s3>=1.24.20
python-box==6.0.2
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
python-multipart>=0.0.5
PyPDF2>=3.0.1
python-docx>=0.8.11
aiofiles>=23.1.0
dvc
-e .
